fix: stop discarding PDF renders when display width is fractional
The render guard compared page.DisplayWidth (double, e.g. 612.7) against
targetWidth ((uint) of the same value, 612), which is false for any
fractional width. Lowering the zoom from fit-width yields fractional
widths, so every rendered bitmap was dropped and pages stuck loading.
Compare the truncated value on both sides.

+++ PreviewWindow.xaml.cs
@@ -840,7 +840,7 @@ namespace Filey
 
                                         Dispatcher.BeginInvoke(new Action(() =>
                                         {
-                                            if (!scrollToken.IsCancellationRequested && !_pdfCts.IsCancellationRequested && page.DisplayWidth == targetWidth)
+                                            if (!scrollToken.IsCancellationRequested && !_pdfCts.IsCancellationRequested && (uint)page.DisplayWidth == targetWidth)
                                             {
                                                 page.Image = bitmap;
                                             }
